facilitator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
cdp-sdk>=0.15.0
python-dotenv>=1.0.0
eth-account>=0.13.0
eth-abi>=5.0.0
web3>=7.0.0
httpx>=0.27.0
